requirements-for-training.txt
accelerate
addict
aiohttp
attrdict
binpacking
charset_normalizer
cpm_kernels
dacite
datasets>=3.0,<3.4
einops
fastapi
gradio>=3.40.0
importlib_metadata
jieba
matplotlib
modelscope>=1.23
nltk
numpy<2.0
openai
oss2
pandas
peft>=0.11,<0.16
pillow
requests
rouge
safetensors
sentencepiece
tensorboard
tiktoken
tqdm
jsonlines
transformers_stream_generator
uvicorn
zstandard
sortedcontainers>=1.5.9
qwen-vl-utils==0.0.6
deepspeed==0.16.0
liger-kernel
vllm==0.9.1
blake3
numpy==1.26.4
triton==3.3.0
pydantic==2.11.7
pydantic_core==2.33.2
# flash_attn-2.7.4.post1+cu12torch2.6cxx11abiTRUE-cp310-cp310-linux_x86_64.whl
# trl_0.17.0
environment-modules-customed/transformers_4.51.3-xwt-customed/transformers/
environment-modules-customed/flash_attn-2.7.4.post1+cu12torch2.6cxx11abiTRUE-cp310-cp310-linux_x86_64.whl
environment-modules-customed/trl